+++ Android/UdacityExercises/Lesson09b-ToDo-List-AAC/T09b.08-Exercise-AddLiveDataToAddTaskActivity/app/src/main/java/com/example/android/todolist/database/TaskDao.java
@@ -25,7 +25,6 @@ public interface TaskDao {
     @Delete
     void deleteTask(TaskEntry taskEntry);
 
-    // TODO (1) Wrap the return type with LiveData
     @Query("SELECT * FROM task WHERE id = :id")
-    TaskEntry loadTaskById(int id);
+    LiveData<TaskEntry> loadTaskById(int id);
 }